Preserve case when parsing domain names
When implementing DNS servers, one might need to return
domain names in exactly the same case as was in the question.
For example, `dig eXamPle.cOm.` includes

```
;; ANSWER SECTION:
eXamPle.cOm.		20885	IN	A	93.184.216.34
```

Notice the answer section says `eXamPle.cOm` and not
`example.com`. This is currently not possible when using this
crate as it automatically converts to lowercase.

Fixes https://github.com/LinkTed/dns-message-parser/issues/12.